Conflicts:
	.idea/compiler.xml
	editor/editor-runtime/models/runtime/impl/cellActions.mps
	editor/editor-runtime/source_gen.caches/jetbrains/mps/editor/runtime/impl/cellActions/dependencies
	editor/editor-runtime/source_gen.caches/jetbrains/mps/editor/runtime/impl/cellActions/generated
	editor/editor-runtime/source_gen/jetbrains/mps/editor/runtime/impl/cellActions/CellAction_Comment.java
	editor/editor-runtime/source_gen/jetbrains/mps/editor/runtime/impl/cellActions/Cell_Action_Uncomment.java
	editor/editor-runtime/source_gen/jetbrains/mps/editor/runtime/impl/cellActions/trace.info
	languages/core/core/languageModels/editor.mps
	languages/core/core/source_gen.caches/jetbrains/mps/lang/core/editor/generated
	languages/core/core/source_gen.caches/jetbrains/mps/lang/core/generated
	languages/languageDesign/plugin/generator/source_gen/jetbrains/mps/lang/plugin/generator/baseLanguage/template/main/trace.info
	plugins/debugger-api/runtime/source_gen/jetbrains/mps/debugger/api/ui/actions/trace.info
	plugins/modelchecker/source_gen/jetbrains/mps/ide/modelchecker/actions/trace.info
	plugins/mpsdevkit/languages/pluginSolutions/lang.structure/source_gen/jetbrains/mps/lang/structure/pluginSolution/plugin/trace.info
	plugins/mpsdevkit/source_gen/jetbrains/mps/ide/devkit/actions/trace.info
	plugins/mpsmake/source_gen/jetbrains/mps/ide/make/actions/trace.info
	plugins/vcs/common/jetbrains.mps.ide.vcs.platform.msd
	plugins/vcs/common/models/diff/ui/common.mps
	plugins/vcs/common/source_gen.caches/jetbrains/mps/vcs/diff/ui/common/dependencies
	plugins/vcs/common/source_gen.caches/jetbrains/mps/vcs/diff/ui/common/generated
	plugins/vcs/common/source_gen/jetbrains/mps/vcs/diff/ui/common/trace.info
	plugins/vcs/source_gen/jetbrains/mps/vcs/plugin/trace.info
	testbench/solutions/jetbrains/mps/tests.mps
	testbench/tests_gen.caches/jetbrains/mps/tests/generated
	testbench/testsolutions/bl.test/test_gen.caches/jetbrains/mps/editorTest/dependencies
	testbench/testsolutions/bl.test/test_gen/jetbrains/mps/editorTest/trace.info
	workbench/mps-editor/solution/actions.mps
	workbench/mps-editor/source_gen.caches/jetbrains/mps/ide/editor/actions/dependencies
	workbench/mps-editor/source_gen.caches/jetbrains/mps/ide/editor/actions/generated
	workbench/mps-editor/source_gen/jetbrains/mps/ide/editor/actions/trace.info
	workbench/mps-workbench/jetbrains.mps.ide/source_gen/jetbrains/mps/ide/actions/trace.info
	workbench/mps-workbench/pluginSolutions/baseLanguage.lightweightdsl/source_gen/jetbrains/mps/baseLanguage/lightweightdsl/pluginSolution/plugin/trace.info

+++ core/kernel/source_gen/jetbrains/mps/scope/Scope.java
@@ -81,7 +81,7 @@ public abstract class Scope {
    */
   public static Scope getScope(SNode node, String role, int index, SNode kind) {
     if (SNodeOperations.isInstanceOf(node, MetaAdapterFactory.getInterfaceConcept(0xceab519525ea4f22L, 0x9b92103b95ca8c0cL, 0x33d23ee961a0cbf3L, "jetbrains.mps.lang.core.structure.ScopeProvider"))) {
-      Scope scope = ((Scope) BHReflection.invoke(SNodeOperations.cast(node, MetaAdapterFactory.getInterfaceConcept(0xceab519525ea4f22L, 0x9b92103b95ca8c0cL, 0x33d23ee961a0cbf3L, "jetbrains.mps.lang.core.structure.ScopeProvider")), SMethodTrimmedId.create("getScope", null, "6GEzh_Hz_wK"), kind, role, index));
+      Scope scope = ((Scope) BHReflection.invoke(SNodeOperations.cast(node, MetaAdapterFactory.getInterfaceConcept(0xceab519525ea4f22L, 0x9b92103b95ca8c0cL, 0x33d23ee961a0cbf3L, "jetbrains.mps.lang.core.structure.ScopeProvider")), SMethodTrimmedId.create("getScope", null, "6GEzh_Hz_wK"), kind, role, ((int) index)));
       if (scope != null) {
         return scope;
       }
